+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Update testing/linting configuration.
 * Add testing for `utils` methods.
 * Add `mypy` for type checking during development.
+* Add `pyonms.utils.check_ip_address()` to validate IP addresses.
 
 **Full Changelog**: https://github.com/mmahacek/PyONMS/compare/v0.1.2...v0.1.3